First iteration with creature render
We have implemented the first iteration rendering a character. We've
created some components to render the different body parts of the
character and accepts an index. This index will determine what image is
load for that body part, since images will be uploaded with a number
name (starting in 0).

Also, we added a library that helps with the spritesheet animation. It
was very simple to animate the spritesheet using this library.